slack/requirements.txt
slack_bolt
requests
pyyaml
